+++ src/test/java/com/tobe/healthy/schedule/ScheduleTest.java
@@ -6,7 +6,7 @@ import static com.tobe.healthy.member.domain.entity.MemberCategory.TRAINER;
 import static java.time.LocalDateTime.of;
 
 import com.tobe.healthy.member.domain.entity.Member;
-import com.tobe.healthy.schedule.application.ScheduleCommandResponse;
+import com.tobe.healthy.schedule.domain.dto.out.ScheduleCommandResponse;
 import com.tobe.healthy.schedule.application.ScheduleService;
 import com.tobe.healthy.schedule.domain.dto.in.AutoCreateScheduleCommandRequest;
 import com.tobe.healthy.schedule.domain.dto.out.ScheduleCommandResult;
